+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- skipping autocomplete of commands when current line is not empty
+
 ## [v0.1.3] - 2020-11-03
 
 ### Added